Moved api module to utils directory
